add delaycompress for elsa logs
When a mailer is installed, occasionally, cron will notify the root user with the cryptic error:
  gzip: stdin: file size changed while zipping
Running logrotate in debug mode identifies the node.log file as the culprit.  Adding the "delaycompress" directive to the logrotate configuration for that file delays the compresson of that file until the next rotation cycle, preventing the email from cron.